Claude/profile progress UI overhaul 011 c ux zg1fui fq7n8 si ezmys

+++ lib/presentation/widgets/progress/before_after_comparison.dart
@@ -179,7 +179,7 @@ class BeforeAfterComparison extends StatelessWidget {
     required String title,
     required String date,
     required double bioAge,
-    required int chronologicalAge,
+    required double chronologicalAge,
     required Color color,
   }) {
     final difference = bioAge - chronologicalAge;
